perf: faster legacy table chart (#385)
* feat: faster legacy table chart

This commit tries to optimize the performance of the legacy data table.

1. Converting everything to Typescript.
2. Create a native React component instead of `reactify` (although all
   DOM operaions still happen in the jQuery.DataTables plugin.
3. Remove dependency on d3, optimize how bars are rendered.

Plus some minor changes to fix linting and building.

Also added a script to build only specific plugin for faster
development.

* feat(legacy-table-chart): use React to render DOM

Unfortunately jquery.datatables uses innerHTML to create cell content,
and all rows are created even with pagination.

https://github.com/DataTables/DataTables/blob/83657a29e33ce93ee940ce25684940eb3acb2913/media/js/jquery.dataTables.js#L3113-L3117

This is slow and insecure. We are reverting to DOM data source as in
previous implementation, but instead of using D3 to create the table rows,
we use React. This also renders `dompurify.sanitize` unnecessary since
React will take care of it.

* feat(legacy-table-chart): support html in data cells

Also

1. improve height adjustment
2. add page size control to storybook
3. hide rows from later pages at initial rendering

* fix(legacy-data-table): minor formatting fixes

* chore(legacy-table): add xss dependency

Plus minor variable name and comment update

* fix(legacy-table): linting errors

* refactor(legacy-table): more predictable metric labels

* feat(legacy-table): also display title for percent metric

* fix(legacy-table): typos, var names, etc

* docs: update notes for metric.label

* refactor(legacy-table): upgrade number-format

* fix(legacy-table): upgrade dependency for storybook and yarn